+++ config/api/item/parameters.php
@@ -12,6 +12,20 @@ return [
             "type" => "string",
             "required" => false
         ],
+        'include-categories' => [
+            'field' => 'include-categories',
+            'title' => 'resource-type-item/parameters.title-include-categories',
+            'description' => 'resource-type-item/parameters.description-include-categories',
+            'type' => 'boolean',
+            'required' => false
+        ],
+        'include-subcategories' => [
+            'field' => 'include-subcategories',
+            'title' => 'resource-type-item/parameters.title-include-subcategories',
+            'description' => 'resource-type-item/parameters.description-include-subcategories',
+            'type' => 'boolean',
+            'required' => false
+        ],
         'year' => [
             "parameter" => "year",
             "title" => 'item/parameters.title-year',